@@ -1,30 +1,33 @@
-from typing import List
-
-from clerk import types
-from clerk.client import Service
+from . import types
+from .client import Service
 
 
 class UsersService(Service):
     endpoint = "users"
 
-    async def list(self) -> List[types.User]:
+    async def list(self) -> list[types.User]:
         """Retrieve a list of all users"""
         async with self._client.get(self.endpoint) as r:
-            return [types.User.parse_obj(s) for s in await r.json()]
+            data = await r.json()
+            print(data)
+            print(type(data))
+            return [types.User.model_validate(s) for s in data]
 
     async def get(self, user_id: str) -> types.User:
         """Retrieve a user by their id"""
         async with self._client.get(f"{self.endpoint}/{user_id}") as r:
-            return types.User.parse_obj(await r.json())
+            return types.User.model_validate(await r.json())
 
     async def delete(self, user_id: str) -> types.DeleteUserResponse:
         """Delete a user by their id"""
         async with self._client.delete(f"{self.endpoint}/{user_id}") as r:
-            return types.DeleteUserResponse.parse_obj(await r.json())
+            return types.DeleteUserResponse.model_validate(await r.json())
 
-    async def update(self, user_id: str, request: types.UpdateUserRequest) -> types.User:
+    async def update(
+        self, user_id: str, request: types.UpdateUserRequest
+    ) -> types.User:
         """Update a user by their id"""
         async with self._client.patch(
             f"{self.endpoint}/{user_id}", data=request.json(exclude_unset=True)
         ) as r:
-            return types.User.parse_obj(await r.json())
+            return types.User.model_validate(await r.json())
